make container tools work with golang+openssl
